tests/wpt/web-platform-tests/webmessaging/message-channels/002.html

<!doctype html>
<title>without start()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function(t) {
  var channel = new MessageChannel();
  channel.port1.postMessage(1);
  var i = 0;
  channel.port2.addEventListener('message', function() { i++; }, false);
  setTimeout(t.step_func(function() { assert_equals(i, 0); t.done();}), 50);
});
</script>
